#2 Переделал структуру компонента вывода состояния части тела
Избавился от мэппинга внутри компонента, чтобы избежать костылей
